+++ jpf-domain/src/main/java/com/joiest/jpf/common/po/PayShopCouponActiveExample.java
@@ -896,52 +896,52 @@ public class PayShopCouponActiveExample {
             return (Criteria) this;
         }
 
-        public Criteria andDouEqualTo(BigDecimal value) {
+        public Criteria andDouEqualTo(Integer value) {
             addCriterion("dou =", value, "dou");
             return (Criteria) this;
         }
 
-        public Criteria andDouNotEqualTo(BigDecimal value) {
+        public Criteria andDouNotEqualTo(Integer value) {
             addCriterion("dou <>", value, "dou");
             return (Criteria) this;
         }
 
-        public Criteria andDouGreaterThan(BigDecimal value) {
+        public Criteria andDouGreaterThan(Integer value) {
             addCriterion("dou >", value, "dou");
             return (Criteria) this;
         }
 
-        public Criteria andDouGreaterThanOrEqualTo(BigDecimal value) {
+        public Criteria andDouGreaterThanOrEqualTo(Integer value) {
             addCriterion("dou >=", value, "dou");
             return (Criteria) this;
         }
 
-        public Criteria andDouLessThan(BigDecimal value) {
+        public Criteria andDouLessThan(Integer value) {
             addCriterion("dou <", value, "dou");
             return (Criteria) this;
         }
 
-        public Criteria andDouLessThanOrEqualTo(BigDecimal value) {
+        public Criteria andDouLessThanOrEqualTo(Integer value) {
             addCriterion("dou <=", value, "dou");
             return (Criteria) this;
         }
 
-        public Criteria andDouIn(List<BigDecimal> values) {
+        public Criteria andDouIn(List<Integer> values) {
             addCriterion("dou in", values, "dou");
             return (Criteria) this;
         }
 
-        public Criteria andDouNotIn(List<BigDecimal> values) {
+        public Criteria andDouNotIn(List<Integer> values) {
             addCriterion("dou not in", values, "dou");
             return (Criteria) this;
         }
 
-        public Criteria andDouBetween(BigDecimal value1, BigDecimal value2) {
+        public Criteria andDouBetween(Integer value1, Integer value2) {
             addCriterion("dou between", value1, value2, "dou");
             return (Criteria) this;
         }
 
-        public Criteria andDouNotBetween(BigDecimal value1, BigDecimal value2) {
+        public Criteria andDouNotBetween(Integer value1, Integer value2) {
             addCriterion("dou not between", value1, value2, "dou");
             return (Criteria) this;
         }
